Fixed file paths
When creating the godot project initially, I made it create its own new folder, which led to the creation of duplicate .gitignore and .gitattributes files. I deleted the entire godot project and recreated it in the repository source folder as to avoid this in future, and re-use the existing git files.